cpuquiet: Use device_{on/off}line(dev) function for up/downcore
The device_online(dev) and device_offline(dev) functions will
take care of also updating the CPU device status when we want
to upcore or downcore, meaning that it will sync the status
also on the device API layer, not only to the lower level API
that is related purely to CPU management.

This also ensures that sysfs gets updated accordingly whenever
a change is made by the CPUQuiet framework, used for RQBalance.

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>