drm/msm/sde: fix flush configuration for physical encoder
The sde encoder merges the flush for both controller but
does not configure the split_flush_enable on mdp_top.
This causes stage-1 fault or vsync_failure because right
side lm/sspp are not getting updated on each update. This
patch enables the split_flush only for pingpong split case
and also checks the vsync status on correct register.

Change-Id: I66b3326f342d63dad0b8d78bd12e882accc1dfa9
Signed-off-by: Dhaval Patel <pdhaval@codeaurora.org>
Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>