rqbalance: Run balancing algorithm when CPU usage is high in downcore
When the decider sets the RQBalance state to downcore, we may
still have a very CPU intensive process that is locking CPUFreq
to maximum frequency: this will make our frequency-based
algorithm kickstarter to just do nothing, while the userspace
may be spawning lots of other processes, resulting in a complete
hog of the current CPU available resources without being unable
to unlock more.

Of course, directly running a full evaluation right after that
will just waste CPU cycles and resources, so also introduce a
new worker delay variable "recheck_delay" that wants to be "an
high amount of time", preset to 2*down_delay: this will make us
able to wait enough time so that, if we have a single thread
that is very CPU intensive, we give it time to finish its job
without putting even more load on our system.

So:
If we want to forcefully downcore, and
- We haven't onlined all cores
- Our CPU is intensively used (perhaps by just one thread);
Then, run the worker again after recheck_delay millseconds
and prefer doing an upcore full evaluation to eventually
unlock more resources.

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>