PM: cpuidle/suspend: Add s2idle usage and time state attributes
Add a new attribute group called "s2idle" under the sysfs directory
of each cpuidle state that supports the ->enter_s2idle callback
and put two new attributes, "usage" and "time", into that group to
represent the number of times the given state was requested for
suspend-to-idle and the total time spent in suspend-to-idle after
requesting that state, respectively.

That will allow diagnostic information related to suspend-to-idle
to be collected without enabling advanced debug features and
analyzing dmesg output.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>

Conflicts:
	Documentation/ABI/testing/sysfs-devices-system-cpu

this adds upstream documentation from 2.6.24 - 4.17 kernel for
/sys/devices/system/cpu/cpuX/cpuidle

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>