defconfig: switch from cpu_wake_boost to reworked cpu_input_boost
These boost frequencies were determined empirically by finding a
combination of frequencies that could complete UiBench's Invalidate
test with fewer than 0.10% of frames dropped; this gives the EAS
workload,
which can be used with the EAS model for msm8998 to find the best
combination of frequencies capable of completing the wordload that
minimizes power consumption.

Note that these frequencies should be re-assessed whenever there is a
significant change to device performance or scheduling.

Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>
Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>