kernel: Boost all CPUs to the max when userspace launches an app
Zygote is Android's application process, and it only forks when launching
a new app for the first time. To accelerate app launches, boost all CPUs
to the max for 1250 ms when zygote decides to fork.

Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>

Conflicts:
	kernel/fork.c

adding include <linux/cpufreq.h> [memory leak fix which we not really
need; due to CONFIG_CPU_FREQ_STAT]

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>