[eval][exp] cpuidle: Allow menu governor to enter deeper sleep states after some time
On several contemporary Intel systems, we have observed that the idle
power consumption is sometimes significantly too high, e.g. 105 W
instead of 74 W for several seconds.

We tracked this issue down to patterns that confuse the heuristic of the
menu idle governor. If the governor observes several consecutive short
sleeps, it does expect another short sleep duration despite no immediate
timers, sending the CPU into a shallow sleep state. On a dyntick-idle
kernel, there are no means for the core to enter a more efficient sleep
state if the prediction was wrong. Ironically this is particularly
problematic if some cores of the system have very infrequent activity,
i.e. many cores and optimized configuration for low idle power.

The effect, cause, triggers, mitigation technique and verification
thereof are described in detail in the a paper that is pending
publication [1].

Fixing the heuristic to make a better prediction does not seem to be
generally feasible. The following patch addresses this issue by setting
a timer such that when the an expected immediate wake-up fails to
appear, the CPU is woken up to go into a deeper sleep state. If the
heuristic was right, the timer is simply canceled.

Please consider the patch a draft for discussion. We need to address:

  * Avoid programming the fallback timer when the deepest sleep state is
    already chosen.
  * Determine good default values for the introduced configurables. This
    is difficult due to the large variety of system configurations
affected
    by the menu governor. Nevertheless we believe this can be done such
     that many systems benefit without significant degradation in any
case.
  * Document (or remove) the sysfs entries.

But first, I would like to invite comments if this is going in the right
direction, or if this should be addressed in a different way.

[1]
https://tu-dresden.de/zih/forschung/ressourcen/dateien/projekte/haec/powernightmares.pdf

Signed-off-by: Thomas Ilsche <thomas.ilsche@tu-dresden.de>
Signed-off-by: Marcus Hähnel <mhaehnel@os.inf.tu-dresden.de>
Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>