[eval] msm: kgsl: Increase worker thread priority
Currently, the kgsl worker thread is erroneosly ranked right below
Android's audio threads in terms of priority.

The kgsl worker thread is in the critical path for rendering frames to the
display, so increase its priority to match the priority of MDSS' kthread (mdss_fb0).

Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>

we're already using a WQ_HIGHPRI for kgsl-events, kgsl-mementry &
kgsl-workqueue (+ WQ_UNBOUND) so this needs careful evaluation

further reading:
https://taste.tuxfamily.org/wiki/index.php?title=Tricks_and_tools_for_PREEMPT-RT_kernel
http://subversion.ffado.org/wiki/IrqPriorities

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>