[eval] h930_defconfig: set CONFIG_DEBUG_ALIGN_RODATA as well:
If this option is enabled, sections that may potentially be marked as
read only or non-executable will be aligned up to the section size of
the kernel. This prevents sections from being split into pages and
avoids a potential TLB penalty. The downside is an increase in alignment
and potentially wasted space. Turn on this option if performance is more
important than memory pressure.

sultanxda judged the following:

"Align linker sections"

"The kernel is a relatively small program so the memory penalty from
this is not very significant. Doing this increases performance."

well, so let's roll with it and see how things go ...

Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>